Fixes Custom Auth Provider.
It looks like `PasswordBrokerManager.php` is not finding custom
Authentication user providers. It looks like on line 11 it uses `use
CreatesUserProviders` then on line 73 it creates a new `PasswordBroker`.
This issue arrises because `PasswordBrokerManager` is using a trait and
referring to an array on the trait that is instantiated on
`AuthManager`. Since it is a trait it is creating a brand new empty
array on `PasswordBrokerManager` which causes issues. I am not familiar
enough with the framework to know if this is the best way to fix this,
but it works and will at least get you pointed in the right direction.

Signed-off-by: crynobone <crynobone@gmail.com>